\encoding{UTF-8}
\name{print.MUS.evaluation.result}
\alias{print.MUS.evaluation.result}
\title{Pretty and comprehensive printing of MUS evaluation results}
\description{
  Pretty and comprehensive printing of MUS evaluation results that can
  be used for working papers.
}
\usage{
\method{print}{MUS.evaluation.result}(x, ...)
}
\arguments{
  \item{x}{A MUS.evaluation.result object that you got by executing the
    function MUS.evaluation.}
  \item{...}{Further arguments, currently ignored.}
}
\author{Henning Prömpers <henning@proempers.net>}
\seealso{
  \code{\link{MUS.evaluation}} for evaluation of the extracted and
  audited sample.
}
\keyword{MUS}
\keyword{audit}
